requirements.txt
pandas
numpy
tensorflow[and-cuda]
scikit-learn
imbalanced-learn
pyyaml 
seaborn
matplotlib